requirements.txt
black==24.10.0
boltons @ file:///Users/ec2-user/ci_py311/boltons_1678396074811/work
brotlipy==0.7.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b64zphacdv/croot/certifi_1683875375103/work/certifi
cffi @ file:///Users/ec2-user/ci_py311/cffi_1678315312775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
conda @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_5aeztnykre/croot/conda_1689269911902/work
conda-content-trust @ file:///Users/ec2-user/ci_py311/conda-content-trust_1678409283937/work
conda-libmamba-solver @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b7jucd8kvj/croot/conda-libmamba-solver_1685032330942/work/src
conda-package-handling @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_f61sssj1s3/croot/conda-package-handling_1685024797154/work
conda_package_streaming @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_78ue32m8ay/croot/conda-package-streaming_1685019689310/work
contourpy==1.2.0
cryptography @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d3qwb8u102/croot/cryptography_1686613065937/work
cycler==0.12.1
filelock==3.19.1
fonttools==4.49.0
fsspec==2025.9.0
idna @ file:///Users/ec2-user/ci_py311/idna_1678315819133/work
Jinja2==3.1.6
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
kiwisolver==1.4.5
libmambapy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_91xegtksvi/croot/mamba-split_1685993177539/work/libmambapy
MarkupSafe==3.0.2
matplotlib==3.8.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.5
numpy==1.26.4
packaging @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c3zlhpboab/croot/packaging_1678965318579/work
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
pluggy @ file:///Users/ec2-user/ci_py311/pluggy_1678315400971/work
pycosat @ file:///Users/ec2-user/ci_py311/pycosat_1678378899646/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_19gfn1ib_u/croot/pyopenssl_1678965300171/work
pyparsing==3.1.2
PySocks @ file:///Users/ec2-user/ci_py311/pysocks_1678315868424/work
python-dateutil==2.9.0.post0
pytz==2024.1
requests @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ec2sxfa_zk/croot/requests_1682607519671/work
ruamel.yaml @ file:///Users/ec2-user/ci_py311/ruamel.yaml_1678379291841/work
scikit-learn==1.4.1.post1
scipy==1.12.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.14.0
termcolor==2.4.0
threadpoolctl==3.3.0
toolz @ file:///Users/ec2-user/ci_py311/toolz_1678323009669/work
torch==2.2.2
torchvision==0.17.2
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_03acd6t6ca/croot/tqdm_1679561866522/work
typing_extensions==4.15.0
tzdata==2024.1
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1968zekhce/croot/urllib3_1686163172153/work
zstandard @ file:///Users/ec2-user/ci_py311_2/zstandard_1679338593428/work
